Changed project structure so src folder is located at the same level with .git. Added .gitignore
